smart-waste-classifier/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
